+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Simplify RoomListHandler
+
 ### Deprecated
 
 ### Removed
@@ -34,7 +36,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add AuthenticationApiClient.getSsoUrl
 - Support for Matrix 1.11
-- Allow message sending to be cancelled while the message is sent
 
 ### Changed